backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-dotenv==1.0.1
